refactor conversation service on front-end
